requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
kafka-python==2.0.2
mysql-connector==2.2.9
numpy==1.20.1
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
requests==2.25.1
schedule==1.0.0
six==1.15.0
urllib3==1.26.3
websocket-client==0.58.0
wincertstore==0.2
